renamed prerender-cf-cache-control-construct type script file
